+++ changelog.md
@@ -2,11 +2,14 @@
 
 - Option `insertionPoint` can now accept a DOM node, so styles can no be rendered also in body or even in an iframe.
 
-### Breaking changes
+### Breaking changes for users
 
 - Jss constructor and Jss.setup option `generateClassName` has been removed. A new option is called `createGenerateClassName` which is a factory that returns the old `generateClassName`. We need this to reset counters on SSR for each request.
 
-- Only for plugins: KeyframeRule has been renamed to KeyframesRule and KeyframeRule.type 'keyframe' to 'keyframes'.
+### Breaking changes for plugins
+
+- KeyframeRule has been renamed to KeyframesRule and KeyframeRule.type 'keyframe' to 'keyframes'.
+- RegularRule has been renamed to StyleRule and its type from 'regular' to 'style'.
 
 ## 7.1.7 / 2017-06-15